Refactor structure

+++ druidgame/view/GUI.cpp
@@ -48,16 +48,16 @@ void GUI::DrawBackground(SDL_Renderer* renderer)
 void GUI::DrawHealth(SDL_Renderer* renderer)
 {
 	float healthPercent = (float)Player->GetHealth() / Player->GetMaxHealth();
-	SDL_Rect sRect = { 0, 0, PLAYER_BAR_WIDTH*healthPercent, PLAYER_BAR_HEIGHT };
-	SDL_Rect dRect = { 25*GetScaleX(), 5*GetScaleY(), PLAYER_BAR_WIDTH*healthPercent*GetScaleX(), PLAYER_BAR_HEIGHT*GetScaleY() };
+	SDL_Rect sRect = { 0, 0, (int)(PLAYER_BAR_WIDTH*healthPercent), PLAYER_BAR_HEIGHT };
+	SDL_Rect dRect = { (int)(25*GetScaleX()), (int)(5*GetScaleY()), (int)(PLAYER_BAR_WIDTH*healthPercent*GetScaleX()), (int)(PLAYER_BAR_HEIGHT*GetScaleY()) };
 	SDL_RenderCopy(renderer, Textures[HEALTH_BAR], &sRect, &dRect);
 }
 
 void GUI::DrawStamina(SDL_Renderer* renderer)
 {
 	float staPercent = (float)Player->GetStamina() / Player->GetMaxStamina();
-	SDL_Rect sRect = { 0, 0, PLAYER_BAR_WIDTH*staPercent, PLAYER_BAR_HEIGHT };
-	SDL_Rect dRect = { 25 * GetScaleX(), 25 * GetScaleY(), PLAYER_BAR_WIDTH*staPercent*GetScaleX(), PLAYER_BAR_HEIGHT*GetScaleY() };
+	SDL_Rect sRect = { 0, 0, (int)(PLAYER_BAR_WIDTH*staPercent), PLAYER_BAR_HEIGHT };
+	SDL_Rect dRect = { (int)(25 * GetScaleX()), (int)(25 * GetScaleY()), (int)(PLAYER_BAR_WIDTH*staPercent*GetScaleX()), (int)(PLAYER_BAR_HEIGHT*GetScaleY()) };
 	SDL_RenderCopy(renderer, Textures[STA_BAR], &sRect, &dRect);
 }
